finalização modulo 3

+++ app/Http/Controllers/Admin/PlanController.php
@@ -49,10 +49,19 @@ class PlanController extends Controller
 
     // deletar registros
     public function destroy($url){
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url', $url)
+            ->first();
 
         if(!$plan){
             return redirect()->back();
+        }
+
+        if ($plan->details->count() > 0){
+            return redirect()
+                    ->back()
+                    ->with('error', 'não é possivel deletar pois existem detalhes vinculados');
         }
 
         $plan->delete();
